+++ services/brig/src/Brig/Options.hs
@@ -31,8 +31,9 @@ import Data.Aeson (withText)
 import qualified Data.Aeson as Aeson
 import Data.Aeson.Types (typeMismatch)
 import qualified Data.Char as Char
-import Data.Domain (Domain)
+import Data.Domain (Domain (..))
 import Data.Id
+import Data.Misc (HttpsUrl)
 import Data.Range
 import Data.Scientific (toBoundedInteger)
 import qualified Data.Text as Text
@@ -455,10 +456,19 @@ data Settings = Settings
     -- | When true, search only
     -- returns users from the same team
     setSearchSameTeamOnly :: !(Maybe Bool),
-    -- | When @Nothing@, assume there are no other backends and IDs are always local.
-    -- This means we don't run any queries on federation-related tables and don't
-    -- make any calls to the federator service.
-    setEnableFederationWithDomain :: !(Maybe Domain),
+    -- | FederationDomain is required, even when not wanting to federate with other backends
+    -- (in that case the 'setFederationAllowedDomains' can be set to empty in Federator)
+    -- Federation domain is used to qualify local IDs and handles,
+    -- e.g. 0c4d8944-70fa-480e-a8b7-9d929862d18c@wire.com and somehandle@wire.com.
+    -- It should also match the SRV DNS records under which other wire-server installations can find this backend:
+    --    _wire-server._tcp.<federationDomain>
+    -- Once set, DO NOT change it: if you do, existing users may have a broken experience and/or stop working
+    -- Remember to keep it the same in Galley.
+    -- Example:
+    --   setFederationAllowedDomains:
+    --     - wire.com
+    --     - example.com
+    setFederationDomain :: !(Domain),
     -- | The amount of time in milliseconds to wait after reading from an SQS queue
     -- returns no message, before asking for messages from SQS again.
     -- defaults to 'defSqsThrottleMillis'.
@@ -474,7 +484,12 @@ data Settings = Settings
     -- Customer extensions
 
     -- | Customer extensions.  Read 'CustomerExtensions' docs carefully!
-    setCustomerExtensions :: !(Maybe CustomerExtensions)
+    setCustomerExtensions :: !(Maybe CustomerExtensions),
+    -- | When set; instead of using SRV lookups to discover SFTs the calls
+    -- config will always return this entry. This is useful in Kubernetes
+    -- where SFTs are deployed behind a load-balancer.  In the long-run the SRV
+    -- fetching logic can go away completely
+    setSftStaticUrl :: !(Maybe HttpsUrl)
   }
   deriving (Show, Generic)
 
@@ -596,8 +611,9 @@ Lens.makeLensesFor
     ("setPropertyMaxValueLen", "propertyMaxValueLen"),
     ("setSearchSameTeamOnly", "searchSameTeamOnly"),
     ("setUserMaxPermClients", "userMaxPermClients"),
-    ("setEnableFederationWithDomain", "enableFederationWithDomain"),
-    ("setSqsThrottleMillis", "sqsThrottleMillis")
+    ("setFederationDomain", "federationDomain"),
+    ("setSqsThrottleMillis", "sqsThrottleMillis"),
+    ("setSftStaticUrl", "sftStaticUrl")
   ]
   ''Settings