+++ services/federator/src/Federator/Impl.hs
@@ -32,6 +32,6 @@ import Servant.Mock
 import Servant.Server
 
 app :: Env -> Application
-app _ = serve api (mock api Proxy)
+app _ = serve api (mock api (Proxy @'[]))
   where
     api = Proxy @(ToServantApi API.Api)