+++ services/galley/src/Galley/API/Teams.hs
@@ -52,29 +52,12 @@ module Galley.API.Teams
     canUserJoinTeamH,
     internalDeleteBindingTeamWithOneMemberH,
     internalDeleteBindingTeamWithOneMember,
-    getFeatureStatus,
-    setFeatureStatus,
-    getAllFeaturesH,
-    getSSOStatusInternal,
-    setSSOStatusInternal,
-    getLegalholdStatusInternal,
-    setLegalholdStatusInternal,
-    getTeamSearchVisibilityAvailableInternal,
-    setTeamSearchVisibilityAvailableInternal,
-    getValidateSAMLEmailsInternal,
-    setValidateSAMLEmailsInternal,
-    getDigitalSignaturesInternal,
-    setDigitalSignaturesInternal,
-    getAppLockInternal,
-    setAppLockInternal,
-    DoAuth (..),
   )
 where
 
 import Brig.Types.Team (TeamSize (..))
 import Control.Lens
 import Control.Monad.Catch
-import qualified Data.Aeson as Aeson
 import Data.ByteString.Conversion hiding (fromList)
 import Data.Id
 import qualified Data.Id as Id
@@ -84,13 +67,12 @@ import Data.List1 (list1)
 import Data.Range as Range
 import Data.Set (fromList)
 import qualified Data.Set as Set
-import Data.String.Conversions (cs)
 import Data.Time.Clock (UTCTime (..), getCurrentTime)
 import qualified Data.UUID as UUID
 import qualified Data.UUID.Util as UUID
 import Galley.API.Error as Galley
 import Galley.API.LegalHold
-import qualified Galley.API.TeamNotifications as APITeamQueue
+import qualified Galley.API.Teams.Notifications as APITeamQueue
 import Galley.API.Util
 import Galley.App
 import qualified Galley.Data as Data
@@ -871,122 +853,6 @@ canUserJoinTeam tid = do
       when (size >= limit) $ do
         throwM tooManyTeamMembersOnTeamWithLegalhold
 
-data DoAuth = DoAuth UserId | DontDoAuth
-
-getFeatureStatus ::
-  forall (a :: Public.TeamFeatureName).
-  Public.KnownTeamFeatureName a =>
-  ( TeamId ->
-    Galley (Public.TeamFeatureStatus a)
-  ) ->
-  DoAuth ->
-  TeamId ->
-  Galley (Public.TeamFeatureStatus a)
-getFeatureStatus getter doauth tid = do
-  case doauth of
-    DoAuth uid -> do
-      zusrMembership <- Data.teamMember tid uid
-      void $ permissionCheck (ViewTeamFeature (Public.knownTeamFeatureName @a)) zusrMembership
-    DontDoAuth ->
-      assertTeamExists tid
-  getter tid
-
-setFeatureStatus ::
-  forall (a :: Public.TeamFeatureName).
-  Public.KnownTeamFeatureName a =>
-  ( TeamId ->
-    Public.TeamFeatureStatus a ->
-    Galley (Public.TeamFeatureStatus a)
-  ) ->
-  DoAuth ->
-  TeamId ->
-  Public.TeamFeatureStatus a ->
-  Galley (Public.TeamFeatureStatus a)
-setFeatureStatus setter doauth tid status = do
-  case doauth of
-    DoAuth uid -> do
-      zusrMembership <- Data.teamMember tid uid
-      void $ permissionCheck (ChangeTeamFeature (Public.knownTeamFeatureName @a)) zusrMembership
-    DontDoAuth ->
-      assertTeamExists tid
-  setter tid status
-
-getAllFeaturesH :: UserId ::: TeamId ::: JSON -> Galley Response
-getAllFeaturesH (uid ::: tid ::: _) =
-  json <$> getAllFeatures uid tid
-
-getAllFeatures :: UserId -> TeamId -> Galley Aeson.Value
-getAllFeatures uid tid = do
-  Aeson.object
-    <$> sequence
-      [ getStatus @'Public.TeamFeatureSSO getSSOStatusInternal,
-        getStatus @'Public.TeamFeatureLegalHold getLegalholdStatusInternal,
-        getStatus @'Public.TeamFeatureSearchVisibility getTeamSearchVisibilityAvailableInternal,
-        getStatus @'Public.TeamFeatureValidateSAMLEmails getValidateSAMLEmailsInternal,
-        getStatus @'Public.TeamFeatureDigitalSignatures getDigitalSignaturesInternal,
-        getStatus @'Public.TeamFeatureAppLock getAppLockInternal
-      ]
-  where
-    getStatus ::
-      forall (a :: Public.TeamFeatureName).
-      (Public.KnownTeamFeatureName a, Aeson.ToJSON (Public.TeamFeatureStatus a)) =>
-      (TeamId -> Galley (Public.TeamFeatureStatus a)) ->
-      Galley (Text, Aeson.Value)
-    getStatus getter = do
-      status <- getFeatureStatus @a getter (DoAuth uid) tid
-      let feature = Public.knownTeamFeatureName @a
-      pure $ (cs (toByteString' feature) Aeson..= status)
-
-getSSOStatusInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureSSO)
-getSSOStatusInternal tid = do
-  defStatus <- do
-    featureSSO <- view (options . optSettings . setFeatureFlags . flagSSO)
-    pure $ case featureSSO of
-      FeatureSSOEnabledByDefault -> Public.TeamFeatureStatusNoConfig Public.TeamFeatureEnabled
-      FeatureSSODisabledByDefault -> Public.TeamFeatureStatusNoConfig Public.TeamFeatureDisabled
-  status <- TeamFeatures.getFeatureStatusNoConfig @'Public.TeamFeatureSSO tid
-  pure . fromMaybe defStatus $ status
-
-setSSOStatusInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureSSO) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureSSO)
-setSSOStatusInternal tid status = do
-  let statusValue = Public.tfwoStatus status
-  case statusValue of
-    Public.TeamFeatureDisabled -> throwM disableSsoNotImplemented
-    Public.TeamFeatureEnabled -> pure () -- this one is easy to implement :)
-  TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureSSO tid status
-
-getLegalholdStatusInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureLegalHold)
-getLegalholdStatusInternal tid = do
-  featureLegalHold <- view (options . optSettings . setFeatureFlags . flagLegalHold)
-  case featureLegalHold of
-    FeatureLegalHoldDisabledByDefault -> do
-      let defaultStatus = Public.TeamFeatureStatusNoConfig Public.TeamFeatureDisabled
-      status <- TeamFeatures.getFeatureStatusNoConfig @'Public.TeamFeatureLegalHold tid
-      pure (fromMaybe defaultStatus status)
-    FeatureLegalHoldDisabledPermanently -> do
-      pure (Public.TeamFeatureStatusNoConfig Public.TeamFeatureDisabled)
-
-setLegalholdStatusInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureLegalHold) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureLegalHold)
-setLegalholdStatusInternal tid status@(Public.tfwoStatus -> statusValue) = do
-  do
-    featureLegalHold <- view (options . optSettings . setFeatureFlags . flagLegalHold)
-    case featureLegalHold of
-      FeatureLegalHoldDisabledByDefault -> do
-        pure ()
-      FeatureLegalHoldDisabledPermanently -> do
-        throwM legalHoldFeatureFlagNotEnabled
-  case statusValue of
-    Public.TeamFeatureDisabled -> removeSettings' tid
-    -- FUTUREWORK: We cannot enable legalhold on large teams right now
-    Public.TeamFeatureEnabled -> checkTeamSize
-  TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureLegalHold tid status
-  where
-    checkTeamSize = do
-      (TeamSize size) <- BrigTeam.getSize tid
-      limit <- fromIntegral . fromRange <$> fanoutLimit
-      when (size > limit) $ do
-        throwM cannotEnableLegalHoldServiceLargeTeam
-
 getTeamSearchVisibilityAvailableInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureSearchVisibility)
 getTeamSearchVisibilityAvailableInternal tid = do
   -- TODO: This is just redundant given there is a decent default
@@ -999,56 +865,10 @@ getTeamSearchVisibilityAvailableInternal tid = do
   fromMaybe defConfig
     <$> TeamFeatures.getFeatureStatusNoConfig @'Public.TeamFeatureSearchVisibility tid
 
-setTeamSearchVisibilityAvailableInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureSearchVisibility) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureSearchVisibility)
-setTeamSearchVisibilityAvailableInternal tid status@(Public.tfwoStatus -> statusValue) = do
-  case statusValue of
-    Public.TeamFeatureDisabled -> SearchVisibilityData.resetSearchVisibility tid
-    Public.TeamFeatureEnabled -> pure () -- This allows the option to be set at the team level
-  TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureSearchVisibility tid status
-
-getValidateSAMLEmailsInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureValidateSAMLEmails)
-getValidateSAMLEmailsInternal tid = do
-  -- FUTUREWORK: we may also want to get a default from the server config file here, like for
-  -- sso, and team search visibility.
-  let defaultStatus = Public.TeamFeatureStatusNoConfig Public.TeamFeatureDisabled
-  fromMaybe defaultStatus
-    <$> TeamFeatures.getFeatureStatusNoConfig @'Public.TeamFeatureValidateSAMLEmails tid
-
-setValidateSAMLEmailsInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureValidateSAMLEmails) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureValidateSAMLEmails)
-setValidateSAMLEmailsInternal tid =
-  TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureValidateSAMLEmails tid
-
-getDigitalSignaturesInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureDigitalSignatures)
-getDigitalSignaturesInternal tid = do
-  -- FUTUREWORK: we may also want to get a default from the server config file here, like for
-  -- sso, and team search visibility.
-  let defaultStatus = Public.TeamFeatureStatusNoConfig Public.TeamFeatureDisabled
-  fromMaybe defaultStatus
-    <$> TeamFeatures.getFeatureStatusNoConfig @'Public.TeamFeatureDigitalSignatures tid
-
-setDigitalSignaturesInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureDigitalSignatures) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureDigitalSignatures)
-setDigitalSignaturesInternal tid =
-  TeamFeatures.setFeatureStatusNoConfig @'Public.TeamFeatureDigitalSignatures tid
-
 -- | Modify and get visibility type for a team (internal, no user permission checks)
 getSearchVisibilityInternalH :: TeamId ::: JSON -> Galley Response
 getSearchVisibilityInternalH (tid ::: _) =
   json <$> getSearchVisibilityInternal tid
-
-getAppLockInternal :: TeamId -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureAppLock)
-getAppLockInternal tid = do
-  let defaultStatus =
-        Public.TeamFeatureStatusWithConfig
-          Public.TeamFeatureEnabled
-          (Public.TeamFeatureAppLockConfig (Public.EnforceAppLock False) 60)
-  status <- TeamFeatures.getApplockFeatureStatus tid
-  pure $ fromMaybe defaultStatus status
-
-setAppLockInternal :: TeamId -> (Public.TeamFeatureStatus 'Public.TeamFeatureAppLock) -> Galley (Public.TeamFeatureStatus 'Public.TeamFeatureAppLock)
-setAppLockInternal tid status = do
-  when (Public.applockInactivityTimeoutSecs (Public.tfwcConfig status) < 30) $
-    throwM inactivityTimeoutTooLow
-  TeamFeatures.setApplockFeatureStatus tid status
 
 getSearchVisibilityInternal :: TeamId -> Galley TeamSearchVisibilityView
 getSearchVisibilityInternal = fmap TeamSearchVisibilityView . SearchVisibilityData.getSearchVisibility