+++ services/spar/test-scim-suite/shell.nix
@@ -1,0 +1,5 @@
+{ pkgs ? import (import ../../../nix/sources.nix).nixpkgs {} }:
+with pkgs;
+mkShell {
+  buildInputs = [newman jq coreutils curl];
+}