+++ tools/stern/src/Stern/Intra.hs
@@ -41,6 +41,8 @@ module Stern.Intra
     isBlacklisted,
     setBlacklistStatus,
     getTeamFeatureFlag,
+    getTeamFeatureFlagNoConfig,
+    setTeamFeatureFlagNoConfig,
     setTeamFeatureFlag,
     getSearchVisibility,
     setSearchVisibility,
@@ -439,8 +441,11 @@ getTeamFeatureFlag tid = do
   let req =
         method GET
           . paths ["/i/teams", toByteString' tid, "features", toByteString' (Public.knownTeamFeatureName @a)]
-          . expect2xx
-  responseJsonUnsafe @(Public.TeamFeatureStatus a) <$> catchRpcErrors (rpc' "galley" gly req)
+  resp <- catchRpcErrors $ rpc' "galley" gly req
+  case Bilge.statusCode resp of
+    200 -> pure $ responseJsonUnsafe @(Public.TeamFeatureStatus a) resp
+    404 -> throwE (Error status404 "bad-upstream" "team doesnt exist")
+    _ -> throwE (Error status502 "bad-upstream" "bad response")
 
 setTeamFeatureFlag ::
   forall (a :: Public.TeamFeatureName).
@@ -461,6 +466,44 @@ setTeamFeatureFlag tid status = do
   resp <- catchRpcErrors $ rpc' "galley" gly req
   case statusCode resp of
     200 -> pure ()
+    404 -> throwE (Error status404 "bad-upstream" "team doesnt exist")
+    _ -> throwE (Error status502 "bad-upstream" "bad response")
+
+getTeamFeatureFlagNoConfig ::
+  TeamId ->
+  Public.TeamFeatureName ->
+  Handler Public.TeamFeatureStatusNoConfig
+getTeamFeatureFlagNoConfig tid featureName = do
+  info $ msg "Getting team feature status"
+  gly <- view galley
+  let req =
+        method GET
+          . paths ["/i/teams", toByteString' tid, "features", toByteString' featureName]
+          . expect2xx
+
+  resp <- catchRpcErrors (rpc' "galley" gly req)
+  case statusCode resp of
+    200 -> pure $ responseJsonUnsafe @Public.TeamFeatureStatusNoConfig resp
+    404 -> throwE (Error status404 "bad-upstream" "team doesnt exist")
+    _ -> throwE (Error status502 "bad-upstream" "bad response")
+
+setTeamFeatureFlagNoConfig ::
+  TeamId ->
+  Public.TeamFeatureName ->
+  Public.TeamFeatureStatusValue ->
+  Handler ()
+setTeamFeatureFlagNoConfig tid featureName statusValue = do
+  info $ msg "Setting team feature status for feature without config"
+  gly <- view galley
+  let req =
+        method PUT
+          . paths ["/i/teams", toByteString' tid, "features", toByteString' featureName]
+          . Bilge.json (Public.TeamFeatureStatusNoConfig statusValue)
+          . contentJson
+  resp <- catchRpcErrors $ rpc' "galley" gly req
+  case statusCode resp of
+    200 -> pure ()
+    404 -> throwE (Error status404 "bad-upstream" "team doesnt exist")
     _ -> throwE $ responseJsonUnsafe resp
 
 getSearchVisibility :: TeamId -> Handler TeamSearchVisibilityView